feat: Add Open Telemetric semantic standard compliant log
Updated logApiResponse and logApiError to emit a log that complies with the
standard defined at https://github.com/open-telemetry/semantic-conventions/blob/main/docs/gen-ai/gen-ai-events.md.

The initial issue described doing this in logUserPrompt and logApiResponse
however after digging deeper, it seems like
`event.gen_ai.client.inference.operation.details` is meant to hold the details
of a completed operation, so this change adds them when an API request completes,
successfully or by erroring.

There are several things to follow up on afterwards:
* Handling logs that are too large #11845
* The format of the output messages, input messages, and system instructions
  is correct BUT JSON stringified instead of logging actual full JSON objects.
  I'm not sure if the latter is possible but I think it's what the spec wants.
  No matter what I tried (both fancy conversion functions and then eventually
  JSON.parse wrapping JSON.stringify), I got an error from the open telemetry
  library: https://github.com/open-telemetry/opentelemetry-js/blob/7e0d45f68271c88487354a4fe5eefe4c654efbfb/packages/opentelemetry-core/src/common/attributes.ts#L36
  (or the same error from a different part of the codebase)
* The server address is more or less hardcoded in the non code-assist cases;
  though it's correct afaik right now, we can't 100% trust it over time. Would
  be better to be able to grab this dynamically
* Aligning the clear cut logs, i.e. adding any data that we added here which we
  might also want there (e.g. topP, topK etc)
* I skipped the tool definitions since it's recommended not to emit them by
  default but we may want to optionally log them

The standard has another log type, `event.gen_ai.evaluation.result` but afaik
we're not doing any kind of evaluations (at runtime) that would make sense to log.

Minor detail: I removed a mock from packages/core/src/telemetry/loggers.test.ts because it didn't seem like there was any reason to mock it when we could use the real thing and get the same results (imo the test was testing less with the mock).

Minor detail: I needed to re-declare EVENT_CHAT_COMPRESSION.
I can't figure out what did it but either a circular dependency or
something about import order was introduced when i updated the api error
logs which caused this import to suddenly be a problem. It's pretty
weird that this metric has the same exact name as the corresponding log
anyway, this seems like an accident - all the other metrics end with a
descriptive word like "latency" or "count" so I'm very suspicous it was
an error to use this string.

Fixes #9215